requirements.txt
ccxt
openai
pandas
schedule
python-dotenv
requests
urllib3>=2.6.3
flask>=3.1.3
PyMySQL>=1.1.2
